feat : 센터 레이아웃 페이지 ui data flow에 맞게 변경 중
